// apps/web/src/api/generated/captain-registrations/captainRegistrationsControllerListRegistrationsSortDirection.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type CaptainRegistrationsControllerListRegistrationsSortDirection = typeof CaptainRegistrationsControllerListRegistrationsSortDirection[keyof typeof CaptainRegistrationsControllerListRegistrationsSortDirection];


export const CaptainRegistrationsControllerListRegistrationsSortDirection = {
  asc: 'asc',
  desc: 'desc',
} as const;