// apps/web/src/api/generated/captain-registrations/captainRegistrationsControllerListRegistrationsStatus.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */

export type CaptainRegistrationsControllerListRegistrationsStatus = typeof CaptainRegistrationsControllerListRegistrationsStatus[keyof typeof CaptainRegistrationsControllerListRegistrationsStatus];


export const CaptainRegistrationsControllerListRegistrationsStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  PENDING_APPROVAL: 'PENDING_APPROVAL',
  CONFIRMED: 'CONFIRMED',
  REJECTED: 'REJECTED',
  WAITLISTED: 'WAITLISTED',
  WITHDRAWN: 'WITHDRAWN',
  CHECKED_IN: 'CHECKED_IN',
  DISQUALIFIED: 'DISQUALIFIED',
  REFUND_PENDING: 'REFUND_PENDING',
  REFUNDED: 'REFUNDED',
} as const;