// apps/web/src/api/generated/captain-registrations/captainMatchResponseDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */
import type { CaptainMatchGameResultDto } from './captainMatchGameResultDto';
import type { CaptainMatchOnlineServerDto } from './captainMatchOnlineServerDto';
import type { CaptainMatchOnsiteAssignmentDto } from './captainMatchOnsiteAssignmentDto';
import type { CaptainMatchOpponentDto } from './captainMatchOpponentDto';
import type { CaptainMatchResponseDtoDisputeStatus } from './captainMatchResponseDtoDisputeStatus';
import type { CaptainMatchResponseDtoForfeitStatus } from './captainMatchResponseDtoForfeitStatus';
import type { CaptainMatchResponseDtoOfficialResultStatus } from './captainMatchResponseDtoOfficialResultStatus';
import type { CaptainMatchResponseDtoStatus } from './captainMatchResponseDtoStatus';
import type { CaptainMatchTournamentDto } from './captainMatchTournamentDto';

export interface CaptainMatchResponseDto {
  id: string;
  tournament: CaptainMatchTournamentDto;
  stage: string;
  round: number;
  /** @nullable */
  bracketPosition?: string | null;
  /** @nullable */
  opponent?: CaptainMatchOpponentDto | null;
  /** @nullable */
  scheduledAt?: string | null;
  timezone: string;
  bestOf: number;
  status: CaptainMatchResponseDtoStatus;
  /** @nullable */
  captainTeamScore?: number | null;
  /** @nullable */
  opponentScore?: number | null;
  mapResults: CaptainMatchGameResultDto[];
  /** @nullable */
  winnerTeamId?: string | null;
  forfeitStatus: CaptainMatchResponseDtoForfeitStatus;
  officialResultStatus: CaptainMatchResponseDtoOfficialResultStatus;
  evidenceAvailable: boolean;
  disputeStatus: CaptainMatchResponseDtoDisputeStatus;
  /** @nullable */
  onlineServer?: CaptainMatchOnlineServerDto | null;
  /** @nullable */
  onsiteAssignment?: CaptainMatchOnsiteAssignmentDto | null;
}