// apps/web/src/api/generated/captain-registrations/captainRegistrationListItemDto.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * CLUTCHA API
 * API documentation for the CLUTCHA online and on-site esports tournament platform.
 * OpenAPI spec version: 1.0.0
 */
import type { CaptainRegistrationListItemDtoApprovalStatus } from './captainRegistrationListItemDtoApprovalStatus';
import type { CaptainRegistrationListItemDtoNextAction } from './captainRegistrationListItemDtoNextAction';
import type { CaptainRegistrationListItemDtoPaymentStatus } from './captainRegistrationListItemDtoPaymentStatus';
import type { CaptainRegistrationListItemDtoStatus } from './captainRegistrationListItemDtoStatus';
import type { CaptainRegistrationTournamentSummaryDto } from './captainRegistrationTournamentSummaryDto';

export interface CaptainRegistrationListItemDto {
  registrationId: string;
  tournament: CaptainRegistrationTournamentSummaryDto;
  status: CaptainRegistrationListItemDtoStatus;
  paymentStatus: CaptainRegistrationListItemDtoPaymentStatus;
  approvalStatus: CaptainRegistrationListItemDtoApprovalStatus;
  submittedAt: string;
  /** @nullable */
  rejectionReason?: string | null;
  nextAction: CaptainRegistrationListItemDtoNextAction;
}